protect resolvers with isAuth check

+++ graphql/resolvers/booking.js
@@ -3,19 +3,27 @@ const Event = require('../../models/event');
 const { transformBooking, transformEvent } = require('./merge');
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
     try {
+      // Allow only authenticated users for this resolver
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const bookings = await Booking.find();
       return bookings.map(booking => transformBooking(booking));
     } catch(err) {
       throw err;
     }
   },
-  bookEvent: async args => {
+  bookEvent: async (args, req) => {
     try {
+      // Allow only authenticated users for this resolver
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const fetchedEvent = await Event.findOne({ _id: args.eventId });
       const booking = new Booking({
-        user: '5e868aab8197c91a08815bfa',
+        user: req.userId,
         event: fetchedEvent
       });
       const result = await booking.save();
@@ -24,8 +32,12 @@ module.exports = {
       throw err; 
     }
   },
-  cancelBooking: async args => {
+  cancelBooking: async (args, req) => {
     try {
+      // Allow only authenticated users for this resolver
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const booking = await Booking.findById(args.bookingId).populate('event');
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
